Implement Session's timeout setting & improve docstring.
Session had a timeout arg but, until now, never used it. For Python 2.6 and
newer this is handled by the standard httplib module; older Pythons rely on a
hack in the couchdb.http module.

Right now, the timeout is either None or a time in seconds and therefore
*always* overrides the socket global default timeout. The Session docstring was
improved slightly to make that clearer.

This commit was heavily based on a patch by daevaorn (thanks!) but greatly
simplified to keep the Python version -specific code out of the "real"
couchdb-python code.