Fix incorrect view aggregation in sync_many (fixes issue 218).
Beforehand, ViewDefinition.sync_many did not sort the views collection which
caused the itertools.groupby call to not aggregates views in the expected
manner (see Python documentation for itertools.groupby for why it is essential
to sort the collection before the groupby call). This would cause conflicts if
a design document contained multiple views, and would cause syncing failures
(such as views not existing upon initial sync).